Change pool check/setItem methods to use size_t
Original API was defined in terms of uint, however, tests use it as if
intention was to support size_t.

+++ src/ocean/util/container/pool/FreeList.d
@@ -262,13 +262,13 @@ version ( UnitTest )
             return new char[10];
         }
 
-        protected override void setItem ( ref Item item, int i )
+        protected override void setItem ( ref Item item, size_t i )
         {
             item.length = 1;
             item[0] = cast(char)(i + 32);
         }
 
-        protected override void checkItem ( ref Item item, int i )
+        protected override void checkItem ( ref Item item, size_t i )
         {
             assert(item.length == 1, "item length wrong");
             assert(item[0] == cast(char)(i + 32), "item content wrong");
@@ -284,7 +284,7 @@ version ( UnitTest )
 
     struct Struct
     {
-        int i;
+        size_t i;
         char[] s;
     }
 
@@ -298,14 +298,14 @@ version ( UnitTest )
             return new Struct;
         }
 
-        protected override void setItem ( ref Item item, int i )
+        protected override void setItem ( ref Item item, size_t i )
         {
             item.i = i;
             item.s.length = 1;
             item.s[0] = cast(char)(i + 32);
         }
 
-        protected override void checkItem ( ref Item item, int i )
+        protected override void checkItem ( ref Item item, size_t i )
         {
             assert(item.i == i, "item integer wrong");
             assert(item.s.length == 1, "item string length wrong");
@@ -322,7 +322,7 @@ version ( UnitTest )
 
     class Class
     {
-        int i;
+        size_t i;
         char[] s;
     }
 
@@ -336,14 +336,14 @@ version ( UnitTest )
             return new Class;
         }
 
-        protected override void setItem ( ref Item item, int i )
+        protected override void setItem ( ref Item item, size_t i )
         {
             item.i = i;
             item.s.length = 1;
             item.s[0] = cast(char)(i + 32);
         }
 
-        protected override void checkItem ( ref Item item, int i )
+        protected override void checkItem ( ref Item item, size_t i )
         {
             assert(item.i == i, "item integer wrong");
             assert(item.s.length == 1, "item string length wrong");
